собрал, надо навести красоту

+++ src/main/java/ru/kata/spring/boot_security/demo/model/Role.java
@@ -1,9 +1,14 @@
 package ru.kata.spring.boot_security.demo.model;
 
-import lombok.*;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
 
 import javax.persistence.*;
+import java.util.Set;
 
 @Entity
 @Table(name = "roles")
@@ -15,7 +20,7 @@ public class Role implements GrantedAuthority {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     @Column(name = "name", unique = true, length = 100)
     private String name;
@@ -29,5 +34,4 @@ public class Role implements GrantedAuthority {
     public String toString() {
         return name.replace("ROLE_", "");
     }
-
 }
